+++ tests/testthat/test-sparse-casters.R
@@ -69,7 +69,6 @@ test_that("Can cast tables into a sparse TermDocumentMatrix", {
 test_that("Can cast tables into a sparse dfm", {
   library(methods)
   skip_if_not_installed("quanteda")
-  data("data_corpus_inaugural", package = "quanteda")
 
   d <- cast_dfm(dat, a, b, val)
   d2 <- cast_dfm_(dat, "a", "b", "val")