+++ NEWS.md
@@ -1,6 +1,7 @@
 #tidytext 0.1.5 [WIP]
 
 * Change how `NA` values are handled in `unnest_tokens` so they no longer cause other columns to become `NA` (#82).
+* Update tidiers and casters to align with quanteda v1.0 (#87).
 
 # tidytext 0.1.4